+++ src/com/kaylerrenslow/armaDialogCreator/arma/control/impl/BasicTextRenderer.java
@@ -107,7 +107,7 @@ public class BasicTextRenderer {
 				return renderer.getLeftX() + (int) (renderer.getWidth() * 0.01);
 			}
 			case ALIGN_RIGHT: {
-				return renderer.getRightX() - (int) (renderer.getWidth() * 0.01);
+				return renderer.getRightX() - textWidth - (int) (renderer.getWidth() * 0.01);
 			}
 			default:
 			case ALIGN_CENTER: {
@@ -122,11 +122,13 @@ public class BasicTextRenderer {
 	}
 
 	public void paint(GraphicsContext gc) {
+		gc.beginPath();
 		gc.rect(renderer.getLeftX(), renderer.getTopY(), renderer.getWidth(), renderer.getHeight());
 		gc.clip();
 		gc.setFont(getFont());
 		gc.setFill(textColor);
 		gc.fillText(getText(), getTextX(), getTextY());
+		gc.closePath();
 	}
 
 	private Font getFont() {